* Overhauled the player physics. Still not perfect, but I'm getting there... * Default blocks can be created with "E" and deleted with "Q". The player's viewing direction is used to dertermine the targer position * Added some place holder trees * Moved "jumping" to "KEY_SPACE" * Added a random number generator the perlin noise generator * Added a seed variable to the world class, which is used for the tree generation and the perlin noise * Optimized the rendering of the display lists * Added the possibility to render translucent blocks (e.g. leafs) * ...

+++ src/blockmania/RenderObject.java
@@ -30,7 +30,7 @@ public abstract class RenderObject {
     public void render() {
     }
 
-    public void update(int delta) {
+    public void update(long delta) {
         
     }
